refactor : create global component folder
