+++ src/routers/CommonRouter.tsx
@@ -1,7 +1,7 @@
 import { Outlet } from "react-router-dom";
 
-import Loading from "@components/Loading";
-import Toast from "@components/Toast";
+import Loading from "@components/global/Loading";
+import Toast from "@components/global/Toast";
 
 const CommonRouter = () => {
   return (
